fixed review comments

+++ backend/sites/src/app/services/folio/folioContents.service.ts
@@ -2,8 +2,9 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { FolioContents } from '../../entities/folioContents.entity';
 import { Repository } from 'typeorm';
-import { FolioContentDTO } from '../../dto/folioContent';
+import { FolioContentDTO } from '../../dto/folioContent.dto';
 import { plainToInstance } from 'class-transformer';
+import { elementAt } from 'rxjs';
 
 @Injectable()
 export class FolioContentsService {
@@ -49,32 +50,40 @@ export class FolioContentsService {
         if (result) {
           return true;
         }
+        else
+        {
+          return false;
+        }
       }
     } catch (error) {
-      throw new Error('Failed to add folio content');
+      throw error;
     }
   }
 
 
   async deleteFolioContent(folioContentId: string): Promise<boolean> {
     try {
-      if (folioContentId != '') {
+      if (folioContentId && folioContentId != '') {
         const result = await this.folioContentRepository.delete({
           id: folioContentId,
         });
         if (result.affected > 0) return true;
         else return false;
       }
+      else
+      {
+        console.log("folioContentId is null or empty")
+      }
 
       return false;
-    } catch (e) {
-      return false;
+    } catch (error) {
+      throw error;
     }
   }
 
   async deleteSitesInFolio(folioId: string, siteId: string ): Promise<boolean> {
     try {
-      if (folioId && siteId) {
+      if (folioId && siteId && folioId !== '' && siteId !== '') {
         const result = await this.folioContentRepository.delete({
           folioId: folioId,
           siteId: siteId
@@ -82,27 +91,35 @@ export class FolioContentsService {
         if (result.affected > 0) return true;
         else return false;
       }
+      else
+      {
+        console.error("folioId or siteId is null or empty.")
+      }
 
       return false;
-    } catch (e) {
-      return false;
+    } catch (error) {
+      throw error;
     }
   }
 
   async deleteAllSitesInFolio(folioId: string): Promise<boolean> {
     try {
       
-      if (folioId) {
+      if (folioId && folioId !== '') {
         const result = await this.folioContentRepository.delete({
           folioId: folioId,         
         });
         if (result.affected > 0) return true;
         else return false;
       }
+      else
+      {
+        console.error("folio id is null or empty");
+      }
 
       return false;
-    } catch (e) {
-      return false;
+    } catch (error) {
+      throw error;
     }
   }
 }
